Changed how we store data to just be a sensor with type
